Break out a helper class from GenerateVisitor annotation processor. (#3060)
* Break out a helper class from GenerateVisitor annotation processor.
This allows service loading all annotation processors and even determining that this one does not apply,
without having javapoet in the classpath.
